Prompt to remove files that have been changed on a stage

+++ Lib/BrawlInstallerLib.py
@@ -5801,7 +5801,7 @@ def moveStageFiles(stageParamList, brstmFiles=[]):
 		writeLog("Finished moving stage files")
 
 # Remove stage entry
-def removeStageEntry(stageParams):
+def removeStageEntry(stageParams, deleteParam=True):
 		writeLog("Removing stage entries")
 		title = "Remove file?"
 		removePac = False
@@ -5811,21 +5811,21 @@ def removeStageEntry(stageParams):
 		reviewedFiles = []
 		for stageParam in stageParams:
 			messageText = "Would you like to remove the following file from " + stageParam.aslEntry.Name + "?:\n"
-			if stageParam.originalPacName:
+			if stageParam.originalPacName and stageParam.originalPacName != stageParam.pacName:
 				pacFile = MainForm.BuildPath + '/pf/stage/melee/' + "STG" + stageParam.originalPacName + ".pac"
 				if File.Exists(pacFile) and pacFile not in reviewedFiles:
 					removePac = BrawlAPI.ShowYesNoPrompt(messageText + "STG" + stageParam.originalPacName + ".pac", title)
 					reviewedFiles.append(pacFile)
 				elif pacFile in reviewedFiles:
 					removePac = False
-			if stageParam.originalModule:
+			if stageParam.originalModule and stageParam.originalModule != stageParam.module:
 				moduleFile = MainForm.BuildPath + '/pf/module/' + stageParam.originalModule
 				if File.Exists(moduleFile) and moduleFile not in reviewedFiles:
 					removeModule = BrawlAPI.ShowYesNoPrompt(messageText + stageParam.originalModule, title)
 					reviewedFiles.append(moduleFile)
 				elif moduleFile in reviewedFiles:
 					removeModule = False
-			if stageParam.originalSoundBank:
+			if stageParam.originalSoundBank and stageParam.originalSoundBank != stageParam.soundBank:
 				directory = Directory.CreateDirectory(MainForm.BuildPath + '/pf/sfx')
 				files = directory.GetFiles(addLeadingZeros(str(hexId(stageParam.originalSoundBank)).replace('0x',''), 3) + "*.sawnd")
 				if len(files) > 0:
@@ -5844,10 +5844,11 @@ def removeStageEntry(stageParams):
 				File.Delete(tracklistFile)
 			if removeSoundbank:
 				File.Delete(files[0].FullName)
-			if stageParam.originalName:
-				paramFile = MainForm.BuildPath + '/pf/stage/stageinfo/' + stageParam.originalName + ".param"
-				if File.Exists(paramFile):
-					File.Delete(paramFile)
+			if deleteParam:
+				if stageParam.originalName:
+					paramFile = MainForm.BuildPath + '/pf/stage/stageinfo/' + stageParam.originalName + ".param"
+					if File.Exists(paramFile):
+						File.Delete(paramFile)
 		writeLog("Finished removing stage entries")
